refactor: 패키지 명 변경 및, NcpClients 적용
